Enhancements/Corrections to Last Block to Form and Block Level Logic.
Tagged changes 20200308 (Determine Type of Trading) where logic was
added to check if the current price bar had blocked when determine what
value to use when setting the Last Block to Form value for the current
price bar.  Previously the logic always defaulted to use the Last Block
to Form value from the previous day.